requirements/base.txt
django==5.2
python-dotenv== 1.1.0 

django-autoslug==1.9.9
django-phonenumber-field==8.1.0 
phonenumbers==8.13.32

Pillow==11.2.1 
argon2-cffi==25.1.0 
djoser==2.3.1 
django-taggit==6.1.0 
django-filter==25.1 

django-celery-beat==2.8.1 
redis==6.2.0 
celery==5.5.3 
flower

asgiref==3.8.1 
attrs==25.3.0 
boto3==1.38.35 
botocore==1.38.35 
cffi==1.17.1
cryptography==45.0.4 

django-cors-headers==4.7.0 
django-storages==1.14.6 
djangorestframework==3.16.0 
djangorestframework-simplejwt==5.5.0 
drf-spectacular==0.28.0

